# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/inventory/skus": {
      "get": {
        "operationId": "InventoryController_getSKUs",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InventoryController_createSKU",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkuDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/inventory/skus/{id}": {
      "get": {
        "operationId": "InventoryController_getSKU",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "InventoryController_updateSKU",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "InventoryController_deleteSKU",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/inventory/stock-ledger/{skuId}": {
      "get": {
        "operationId": "InventoryController_getStockLedger",
        "parameters": [
          {
            "name": "skuId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/inventory/stock-ledger": {
      "post": {
        "operationId": "InventoryController_createStockLedgerEntry",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/inventory/warehouses": {
      "get": {
        "operationId": "InventoryController_getWarehouses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InventoryController_createWarehouse",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/inventory/bins": {
      "get": {
        "operationId": "InventoryController_getBins",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InventoryController_createBin",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/inventory/skus/{id}/onhand": {
      "get": {
        "operationId": "InventoryController_getOnHand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "onHand": 42
                  }
                }
              }
            }
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/purchasing": {
      "get": {
        "operationId": "PurchasingController_getAll",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PurchasingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/purchasing/{id}": {
      "get": {
        "operationId": "PurchasingController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/purchasing/{id}/receive": {
      "put": {
        "operationId": "PurchasingController_receive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/purchasing/{id}/cancel": {
      "put": {
        "operationId": "PurchasingController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationController_getNotifications",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/notifications/with-count": {
      "get": {
        "operationId": "NotificationController_getNotificationsWithCount",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/notifications/{id}": {
      "delete": {
        "operationId": "NotificationController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "HVAC API",
    "description": "Generated API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateSkuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreatePoDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "skuId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "skuId",
          "quantity"
        ]
      },
...
    }
  }
}